feat: devided search result in diffrent components for better readablity
