+++ tsconfig.app.json
@@ -1,10 +1,13 @@
 {
   "compilerOptions": {
+    "noEmit": true,
     "allowJs": true,
     "jsx": "react",
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"]
     }
-  }
+  },
+  "include": ["src"],
+  "exclude": ["dist", "node_modules"]
 }
